refactor: student package 분리
