--- backend/src/main/java/com/festago/domain/MailClient.java
@@ -1,6 +1,0 @@
-package com.festago.domain;
-
-public interface MailClient {
-
-    void send(VerificationMailPayload payload);
-}
