--- backend/src/main/java/com/festago/domain/VerificationCodeProvider.java
@@ -1,6 +1,0 @@
-package com.festago.domain;
-
-public interface VerificationCodeProvider {
-
-    VerificationCode provide();
-}
