+++ backend/src/main/java/com/festago/student/application/MailClient.java
@@ -1,0 +1,8 @@
+package com.festago.student.application;
+
+import com.festago.student.domain.VerificationMailPayload;
+
+public interface MailClient {
+
+    void send(VerificationMailPayload payload);
+}
