+++ backend/src/main/java/com/festago/student/application/VerificationCodeProvider.java
@@ -1,0 +1,8 @@
+package com.festago.student.application;
+
+import com.festago.student.domain.VerificationCode;
+
+public interface VerificationCodeProvider {
+
+    VerificationCode provide();
+}
