+++ backend/src/main/java/com/festago/student/domain/VerificationMailPayload.java
@@ -1,4 +1,4 @@
-package com.festago.domain;
+package com.festago.student.domain;
 
 import com.festago.exception.ErrorCode;
 import com.festago.exception.InternalServerException;
