+++ backend/src/main/java/com/festago/student/dto/StudentVerificateRequest.java
@@ -1,4 +1,4 @@
-package com.festago.dto;
+package com.festago.student.dto;
 
 public record StudentVerificateRequest(String code) {
