+++ backend/src/main/java/com/festago/student/infrastructure/GoogleMailClient.java
@@ -1,7 +1,7 @@
-package com.festago.infrastructure;
+package com.festago.student.infrastructure;
 
-import com.festago.domain.MailClient;
-import com.festago.domain.VerificationMailPayload;
+import com.festago.student.application.MailClient;
+import com.festago.student.domain.VerificationMailPayload;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Profile;
 import org.springframework.mail.MailSender;
