+++ backend/src/main/java/com/festago/student/repository/StudentCodeRepository.java
@@ -1,5 +1,8 @@
-package com.festago.domain;
+package com.festago.student.repository;
 
+import com.festago.domain.Member;
+import com.festago.student.domain.StudentCode;
+import com.festago.student.domain.VerificationCode;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
