+++ backend/src/main/java/com/festago/student/repository/StudentRepository.java
@@ -1,5 +1,6 @@
-package com.festago.domain;
+package com.festago.student.repository;
 
+import com.festago.student.domain.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StudentRepository extends JpaRepository<Student, Long> {
