+++ backend/src/test/java/com/festago/application/StudentServiceTest.java
@@ -6,20 +6,21 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.BDDMockito.given;
 
-import com.festago.domain.MailClient;
 import com.festago.domain.Member;
 import com.festago.domain.MemberRepository;
 import com.festago.domain.School;
 import com.festago.domain.SchoolRepository;
-import com.festago.domain.StudentCode;
-import com.festago.domain.StudentCodeRepository;
-import com.festago.domain.StudentRepository;
-import com.festago.domain.VerificationCode;
-import com.festago.domain.VerificationCodeProvider;
-import com.festago.dto.StudentSendMailRequest;
-import com.festago.dto.StudentVerificateRequest;
 import com.festago.exception.BadRequestException;
 import com.festago.exception.NotFoundException;
+import com.festago.student.application.MailClient;
+import com.festago.student.application.StudentService;
+import com.festago.student.application.VerificationCodeProvider;
+import com.festago.student.domain.StudentCode;
+import com.festago.student.domain.VerificationCode;
+import com.festago.student.dto.StudentSendMailRequest;
+import com.festago.student.dto.StudentVerificateRequest;
+import com.festago.student.repository.StudentCodeRepository;
+import com.festago.student.repository.StudentRepository;
 import com.festago.support.MemberFixture;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayNameGeneration;
